snappy: remove autogenerated file
If the intention was to bypass the snappy build system, only one of the
two autogenerated files were checked in, leading to build failure due
to non-existent snappy/config.h.

However, bypassing the snappy build system is categorically incorrect
since it performs CPU feature tests and generates conditional #defines
based on that.